requirements.txt
oracledb==1.9.0
python-dotenv==1.0.0
numpy==1.26.0
Pillow==10.0.0
matplotlib==3.8.0
opencv-python==4.8.0
scikit-image==0.21.0